requirements.txt
slack-bolt==1.18.0
python-dotenv==1.0.0
boto3==1.34.0
